Fix issues with graphql-java 2.0

+++ src/main/java/org/opentripplanner/index/IndexGraphQLSchema.java
@@ -1093,10 +1093,10 @@ public class IndexGraphQLSchema {
                     .build())
                 .dataFetcher(environment -> index.graph.streetIndex
                     .getTransitStopForEnvelope(new Envelope(
-                        new Coordinate((double) (float) environment.getArgument("minLon"),
-                            (double) (float) environment.getArgument("minLat")),
-                        new Coordinate((double) (float) environment.getArgument("maxLon"),
-                            (double) (float) environment.getArgument("maxLat"))))
+                        new Coordinate(environment.getArgument("minLon"),
+                            environment.getArgument("minLat")),
+                        new Coordinate(environment.getArgument("maxLon"),
+                            environment.getArgument("maxLat"))))
                     .stream()
                     .map(TransitVertex::getStop)
                     .filter(stop -> environment.getArgument("agency") == null || stop.getId()
@@ -1132,7 +1132,7 @@ public class IndexGraphQLSchema {
                 .argument(relay.getConnectionFieldArguments())
                 .dataFetcher(environment ->
                     new SimpleListConnection(index.findClosestStopsByWalking(
-                        environment.getArgument("lat"), environment.getArgument("lon"),
+                        (float) (double) environment.getArgument("lat"), (float) (double) environment.getArgument("lon"),
                         environment.getArgument("radius")
                     )
                         .stream()
